Added image preprocessing to `NewTable`
Added image preprocessing to `NewTable` to match the preprocessing done in `FindTable`.
This is done because I suspect that in the future we change the input image format of `NewTable`,
so I did it now. So `FindTable` now returns an RGBA image.
Because of UI reasons we probably want to return the original image with the perspective transform applied to it,
instead of the binary version of the original image.
Because we then take in the normal image in `NewTable` we need to apply the same image manipulations to it as in `FindTable`.