cv.ReviewTable extracts the students' names from the name column.
This is achievied by first preprocessing the image. After that a NewTable applied
on the image, which returns a table based on the image. After that we iterate
through all rows and skip any deformed rows. NewTable already performs some checks
to insure that this doesn't happen, this is just to be safe.
Afterwards be extract the ROI of the name region and convert it to PNG. This is done
because OCRTesseract only accepts images of certain formats like PNG or JPEG.
We use PNG because it is a lossless format and therefore no additional recognition
errors are created through the lossy compression.