Image processing: signature validation
Signature validation was achieved in a very simple way, that should be
subject to change in the future. We do this validation by simply asserting
that each signature field only has one signature in it. This is done by
detecting the contours and getting the bounding boxes of them. We select
the interessting bounding boxes and check that there is only one of them
per signature field.
We should replace signature validation with signature verification where
we also assert that the signature belongs to the correct student.