Added tests for cv.ReviewTable
This test is to ensure that cv.ReviewTable performs all the correct logic.
This is added instead of cv.ValidSignature, because cv.ValidSignature can only be called
in a certain environment, which is set up by cv.ReviewTable. There would be no sense in
recreating that environment in a test function, if it is already created by cv.ReviewTable.
Therefore we simply use cv.ReviewTable.