Added functionality in cv.ValidSignature to merge bounding rectangles
Sometimes it happens, that the contour detection splits a signature apart into two
bounding rectangles. This could then result in the signature field getting marked
as invalid. As an example for where this could happen, take a look at the signature field
in the row designated "11".
To mitigate such issues, I made it, such that after the bounding rectangles were gathered based
on the contours of the signature, those bounding rects were merged together into one rectangle,
which is the bounding box of both former rectangles.
After that the previous algorithm is applied on the merged bounding rectangles.