CV code refactoring: fixed merge, full row bounding rect
I mainly refactored the code in various places to return a more generic type `TableRow`
instead of, for example, `NameROI`. This required also a rewrite of the merge function
where I realised some errors of it and at the end managed to refactor it into a simpler,
but also more readable form.