Deletes all your code, refuses to elaborate further, leaves
As I was interested how it would turn out, I rewrote the database part using GORM as the
ORM of choice. As it turned out This cut down tremendously on the size and therefore
cognitive overload of the database package (also because I did no error handling).
I rewrote some functions which could now be implemented in a much shorter fashion.
Those functions were also made available to the MVVM.